Split save_codec.rs (663 lines) into subdirectory modules
Breaks the monolithic save_codec.rs into four focused submodules:
- grid_codecs.rs: ZoneType, RoadType codecs (56 lines)
- service_codecs.rs: UtilityType, ServiceType codecs (146 lines)
- weather_codecs.rs: Weather, Season, Climate, Fog, Wind, Drought, HeatWave, ColdSnap codecs (188 lines)
- infrastructure_codecs.rs: Policy, UnlockNode, Recycling, Water, Treatment, Compost, Landfill, Reservoir, Sewer codecs (288 lines)

mod.rs re-exports everything via `pub use`, so all external callers
(`use crate::save_codec::*`) continue to work without changes.

Closes #1389
