+++ crates/save/src/save_codec/mod.rs
@@ -1,0 +1,13 @@
+// ---------------------------------------------------------------------------
+// save_codec – Encoding helpers for enum ↔ u8 round-trips
+// ---------------------------------------------------------------------------
+
+mod grid_codecs;
+mod infrastructure_codecs;
+mod service_codecs;
+mod weather_codecs;
+
+pub use grid_codecs::*;
+pub use infrastructure_codecs::*;
+pub use service_codecs::*;
+pub use weather_codecs::*;
