+++ crates/save/src/save_codec/service_codecs.rs
@@ -1,0 +1,146 @@
+// ---------------------------------------------------------------------------
+// Service & utility codecs: UtilityType, ServiceType
+// ---------------------------------------------------------------------------
+
+use simulation::services::ServiceType;
+use simulation::utilities::UtilityType;
+
+pub fn utility_type_to_u8(u: UtilityType) -> u8 {
+    match u {
+        UtilityType::PowerPlant => 0,
+        UtilityType::SolarFarm => 1,
+        UtilityType::WindTurbine => 2,
+        UtilityType::WaterTower => 3,
+        UtilityType::SewagePlant => 4,
+        UtilityType::NuclearPlant => 5,
+        UtilityType::Geothermal => 6,
+        UtilityType::PumpingStation => 7,
+        UtilityType::WaterTreatment => 8,
+    }
+}
+
+pub fn u8_to_utility_type(v: u8) -> UtilityType {
+    match v {
+        0 => UtilityType::PowerPlant,
+        1 => UtilityType::SolarFarm,
+        2 => UtilityType::WindTurbine,
+        3 => UtilityType::WaterTower,
+        4 => UtilityType::SewagePlant,
+        5 => UtilityType::NuclearPlant,
+        6 => UtilityType::Geothermal,
+        7 => UtilityType::PumpingStation,
+        8 => UtilityType::WaterTreatment,
+        _ => UtilityType::PowerPlant, // fallback
+    }
+}
+
+pub fn service_type_to_u8(s: ServiceType) -> u8 {
+    match s {
+        ServiceType::FireStation => 0,
+        ServiceType::PoliceStation => 1,
+        ServiceType::Hospital => 2,
+        ServiceType::ElementarySchool => 3,
+        ServiceType::HighSchool => 4,
+        ServiceType::University => 5,
+        ServiceType::Library => 6,
+        ServiceType::SmallPark => 7,
+        ServiceType::LargePark => 8,
+        ServiceType::Playground => 9,
+        ServiceType::Plaza => 10,
+        ServiceType::SportsField => 11,
+        ServiceType::Stadium => 12,
+        ServiceType::Landfill => 13,
+        ServiceType::RecyclingCenter => 14,
+        ServiceType::Incinerator => 15,
+        ServiceType::Cemetery => 16,
+        ServiceType::Crematorium => 17,
+        ServiceType::CityHall => 18,
+        ServiceType::Museum => 19,
+        ServiceType::Cathedral => 20,
+        ServiceType::TVStation => 21,
+        ServiceType::BusDepot => 22,
+        ServiceType::TrainStation => 23,
+        ServiceType::FireHouse => 24,
+        ServiceType::FireHQ => 25,
+        ServiceType::PoliceKiosk => 26,
+        ServiceType::PoliceHQ => 27,
+        ServiceType::Prison => 28,
+        ServiceType::MedicalClinic => 29,
+        ServiceType::MedicalCenter => 30,
+        ServiceType::Kindergarten => 31,
+        ServiceType::SubwayStation => 32,
+        ServiceType::TramDepot => 33,
+        ServiceType::FerryPier => 34,
+        ServiceType::SmallAirstrip => 35,
+        ServiceType::InternationalAirport => 36,
+        ServiceType::TransferStation => 37,
+        ServiceType::CellTower => 38,
+        ServiceType::DataCenter => 39,
+        ServiceType::HomelessShelter => 40,
+        ServiceType::PostOffice => 41,
+        ServiceType::MailSortingCenter => 42,
+        ServiceType::RegionalAirport => 43,
+        ServiceType::WelfareOffice => 44,
+        ServiceType::HeatingBoiler => 45,
+        ServiceType::DistrictHeatingPlant => 46,
+        ServiceType::GeothermalPlant => 47,
+        ServiceType::WaterTreatmentPlant => 48,
+        ServiceType::WellPump => 49,
+    }
+}
+
+pub fn u8_to_service_type(v: u8) -> Option<ServiceType> {
+    match v {
+        0 => Some(ServiceType::FireStation),
+        1 => Some(ServiceType::PoliceStation),
+        2 => Some(ServiceType::Hospital),
+        3 => Some(ServiceType::ElementarySchool),
+        4 => Some(ServiceType::HighSchool),
+        5 => Some(ServiceType::University),
+        6 => Some(ServiceType::Library),
+        7 => Some(ServiceType::SmallPark),
+        8 => Some(ServiceType::LargePark),
+        9 => Some(ServiceType::Playground),
+        10 => Some(ServiceType::Plaza),
+        11 => Some(ServiceType::SportsField),
+        12 => Some(ServiceType::Stadium),
+        13 => Some(ServiceType::Landfill),
+        14 => Some(ServiceType::RecyclingCenter),
+        15 => Some(ServiceType::Incinerator),
+        16 => Some(ServiceType::Cemetery),
+        17 => Some(ServiceType::Crematorium),
+        18 => Some(ServiceType::CityHall),
+        19 => Some(ServiceType::Museum),
+        20 => Some(ServiceType::Cathedral),
+        21 => Some(ServiceType::TVStation),
+        22 => Some(ServiceType::BusDepot),
+        23 => Some(ServiceType::TrainStation),
+        24 => Some(ServiceType::FireHouse),
+        25 => Some(ServiceType::FireHQ),
+        26 => Some(ServiceType::PoliceKiosk),
+        27 => Some(ServiceType::PoliceHQ),
+        28 => Some(ServiceType::Prison),
+        29 => Some(ServiceType::MedicalClinic),
+        30 => Some(ServiceType::MedicalCenter),
+        31 => Some(ServiceType::Kindergarten),
+        32 => Some(ServiceType::SubwayStation),
+        33 => Some(ServiceType::TramDepot),
+        34 => Some(ServiceType::FerryPier),
+        35 => Some(ServiceType::SmallAirstrip),
+        36 => Some(ServiceType::InternationalAirport),
+        37 => Some(ServiceType::TransferStation),
+        38 => Some(ServiceType::CellTower),
+        39 => Some(ServiceType::DataCenter),
+        40 => Some(ServiceType::HomelessShelter),
+        41 => Some(ServiceType::PostOffice),
+        42 => Some(ServiceType::MailSortingCenter),
+        43 => Some(ServiceType::RegionalAirport),
+        44 => Some(ServiceType::WelfareOffice),
+        45 => Some(ServiceType::HeatingBoiler),
+        46 => Some(ServiceType::DistrictHeatingPlant),
+        47 => Some(ServiceType::GeothermalPlant),
+        48 => Some(ServiceType::WaterTreatmentPlant),
+        49 => Some(ServiceType::WellPump),
+        _ => None,
+    }
+}
